+++ packages/components/src/range-control/types.ts
@@ -116,6 +116,8 @@ export type RangeControlProps = Pick<
 		currentInput?: number;
 		/**
 		 * An icon to be shown above the slider next to its container title.
+		 *
+		 * @deprecated Use `beforeIcon` or `afterIcon` instead.
 		 */
 		icon?: string;
 		/**